respect temperature unit

+++ station/Enums/AlertType.swift
@@ -2,7 +2,7 @@ import Foundation
 
 enum AlertType: CaseIterable {
     
-    case temperature(lower: Int, upper: Int)
+    case temperature(lower: Double, upper: Double)
     
     static var allCases: [AlertType] {
         return [.temperature(lower: 0, upper: 0)]
